+++ packages/sdk/src/endpoints/Endpoint.ts
@@ -53,7 +53,8 @@ export abstract class Endpoint {
                 docs: errorPayload.docs,
                 time: errorPayload.time,
                 code: errorPayload.code,
-                message: errorPayload.message
+                message: errorPayload.message,
+                stacktrace: errorPayload.stacktrace
             });
         }